backend/ml-service/requirements.txt
ultralytics==8.0.208
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
torch==2.0.1
torchvision==0.15.2
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
